modularized program + adding phonne number cleaning logic
