--- v5-src/database.py
@@ -1,294 +1,0 @@
-import sqlite3
-import csv
-
-def clean_data(filename):
-    """
-    reads csv, cleans data, and returns normalized headers, cleaned rows, and a success boolean.
-    """
-    norm_headers = []
-    cleaned_rows = []
-    success = False
-    try:
-        with open(filename, newline='', encoding='utf-8') as csvfile:
-            reader = csv.reader(csvfile)
-            headers = next(reader, None)
-            if not headers:
-                print("[clean_data] error: no headers found in CSV.")
-                return norm_headers, cleaned_rows, success
-            norm_headers = [h.strip().lower().replace(' ', '_') for h in headers]
-            for row in reader:
-                cleaned_row = [cell.strip() if isinstance(cell, str) else '' for cell in row]
-                cleaned_rows.append(cleaned_row)
-        if not cleaned_rows:
-            print("[clean_data] warning: no data rows found in CSV.")
-        else:
-            success = True
-            print(f"[debug] identified headers: {norm_headers}")
-            print(f"[debug] number of data rows: {len(cleaned_rows)}")
-        return norm_headers, cleaned_rows, success
-    except Exception as e:
-        print(f"[clean_data] error: {e}")
-        return norm_headers, cleaned_rows, success
-
-def create_tables(conn, headers, main_table_name):
-    """
-    creates 3nf tables: country, city, location, and a main table (with dynamic columns).
-    returns success boolean.
-    """
-    success = False
-    try:
-        cursor = conn.cursor()
-        print("[debug] creating tables...")
-        cursor.execute('CREATE TABLE IF NOT EXISTS country (country_id INTEGER PRIMARY KEY, country_name TEXT UNIQUE)')
-        cursor.execute('CREATE TABLE IF NOT EXISTS city (city_id INTEGER PRIMARY KEY, city_name TEXT UNIQUE)')
-        cursor.execute('''CREATE TABLE IF NOT EXISTS location (
-            location_id INTEGER PRIMARY KEY,
-            city_id INTEGER,
-            country_id INTEGER,
-            UNIQUE(city_id, country_id),
-            FOREIGN KEY(city_id) REFERENCES city(city_id),
-            FOREIGN KEY(country_id) REFERENCES country(country_id)
-        )''')
-        # main table: all columns except city and country
-        main_headers = [h for h in headers if h not in ('city', 'country')]
-        cols_sql = ', '.join([f'"{col}" TEXT' for col in main_headers])
-        cursor.execute(f'''CREATE TABLE IF NOT EXISTS "{main_table_name}" (
-            {main_table_name}_id INTEGER PRIMARY KEY,
-            location_id INTEGER,
-            {cols_sql},
-            FOREIGN KEY(location_id) REFERENCES location(location_id)
-        )''')
-        print(f"[debug] main table '{main_table_name}' and supporting tables created.")
-        success = True
-    except Exception as e:
-        print(f"[create_tables] error: {e}")
-    return success
-
-def inject(conn, headers, rows, main_table_name):
-    """
-    injects cleaned data into the normalized tables.
-    returns success boolean.
-    """
-    try:
-        cursor = conn.cursor()
-        city_idx = headers.index('city') if 'city' in headers else None
-        country_idx = headers.index('country') if 'country' in headers else None
-        # verify headers for debug
-        if city_idx is None or country_idx is None:
-            print("[inject] error: 'city' or 'country' column not found in headers.")
-            return False
-        main_indices = [i for i in range(len(headers)) if i not in (city_idx, country_idx)]
-        main_headers = [headers[i] for i in main_indices]
-
-        for row in rows:
-            city = row[city_idx]
-            country = row[country_idx]
-            # country
-            cursor.execute('INSERT OR IGNORE INTO country (country_name) VALUES (?)', (country,))
-            cursor.execute('SELECT country_id FROM country WHERE country_name=?', (country,))
-            country_id = cursor.fetchone()[0]
-            # city
-            cursor.execute('INSERT OR IGNORE INTO city (city_name) VALUES (?)', (city,))
-            cursor.execute('SELECT city_id FROM city WHERE city_name=?', (city,))
-            city_id = cursor.fetchone()[0]
-            # location
-            cursor.execute('INSERT OR IGNORE INTO location (city_id, country_id) VALUES (?, ?)', (city_id, country_id))
-            cursor.execute('SELECT location_id FROM location WHERE city_id=? AND country_id=?', (city_id, country_id))
-            location_id = cursor.fetchone()[0]
-            # main table
-            main_values = [row[i] for i in main_indices]
-            placeholders = ', '.join(['?'] * (1 + len(main_headers)))
-            columns_str = ', '.join(['location_id'] + [f'"{col}"' for col in main_headers])            
-            cursor.execute(f'INSERT INTO {main_table_name} ({columns_str}) VALUES ({placeholders})', [location_id] + main_values)
-
-        conn.commit()
-        print(f"[inject] injected {len(rows)} rows into country, city, location, and main tables (3NF).")
-        return True
-    except Exception as e:
-        print(f"[inject] error: {e}")
-        return False
-
-def inject_dataset(conn, filename, main_table_name):
-    """
-    main orchestrator: clean data, create/alter tables, and inject.
-    returns success boolean.
-    """
-    print("[database] inject_dataset called.")
-    print(f"[database] cleaning data from {filename}.")
-    headers, rows, success = clean_data(filename)
-    if not success:
-        print("[database] error: cleaning data failed.")
-        return False
-
-    print("[database] creating tables create_tables(conn, headers, main_table_name)")
-    success = create_tables(conn, headers, main_table_name)
-    if not success:
-        print("[database] error: creating tables failed.")
-        return False
-    
-    print("[database] tables created successfully, now injecting data.")
-
-    print("[database] injecting inject(conn, headers, rows, main_table_name)")
-    success = inject(conn, headers, rows, main_table_name)
-    return success
-
-def drop_all_tables(conn):
-    """
-    drops all user tables in the database for a clean test run.
-    """
-    cursor = conn.cursor()
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
-    tables = [row[0] for row in cursor.fetchall()]
-    for table in tables:
-        cursor.execute(f'DROP TABLE IF EXISTS "{table}"')
-    conn.commit()
-    print("[debug] all tables dropped.")
-
-def print_database_summary(conn, main_table_name):
-    cursor = conn.cursor()
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
-    tables = [row[0] for row in cursor.fetchall()]
-    print("\n[summary] database contains the following tables:")
-    for table in tables:
-        print(f"\n  table: {table}")
-        # get columns
-        cursor.execute(f'PRAGMA table_info("{table}")')
-        columns = [info[1] for info in cursor.fetchall()]
-        print(f"    columns: {columns}")
-        # get row count
-        cursor.execute(f'SELECT COUNT(*) FROM "{table}"')
-        count = cursor.fetchone()[0]
-        print(f"    row count: {count}")
-        # print first 3 rows as a sample
-        cursor.execute(f'SELECT * FROM "{table}" LIMIT 3')
-        rows = cursor.fetchall()
-        if rows:
-            print("    sample rows:")
-            for row in rows:
-                print(f"      {row}")
-        else:
-            print("    (no data)")
-
-    # aggregates from main table
-    print("\n[summary] aggregates from main table:")
-    if main_table_name and main_table_name in tables:
-        # total number of customers (rows in main table)
-        cursor.execute(f'SELECT COUNT(*) FROM "{main_table_name}"')
-        total_customers = cursor.fetchone()[0]
-        print(f"\n[summary] total number of customers: {total_customers}")
-
-        # most common country of origin
-        # joining main table -> location -> country
-        try:
-            cursor.execute(f'''
-                SELECT country.country_name, COUNT(*) as cnt
-                FROM "{main_table_name}"
-                JOIN location ON "{main_table_name}".location_id = location.location_id
-                JOIN country ON location.country_id = country.country_id
-                GROUP BY country.country_name
-                ORDER BY cnt DESC
-                LIMIT 1
-            ''')
-            result = cursor.fetchone()
-            if result:
-                print(f"[summary] most common country of origin: {result[0]} ({result[1]} customers)")
-            else:
-                print("[summary] could not determine most common country.")
-        except Exception as e:
-            print(f"[summary] error aggregating country data: {e}")
-
-def main():
-    import sys
-    import os
-    success = False
-
-    if len(sys.argv) < 2:
-        print("[main] usage: python database.py <csv_filename>")
-        return
-    filename = sys.argv[1]
-    db_path = 'database-content.db'
-    conn = sqlite3.connect(db_path)
-
-    if conn is None:
-        print("[main] error: could not connect to database.")
-        return
-    print(f"[main] connected to database at {db_path}.")
-
-    # drop all tables for a clean test run
-    try:
-        print("[main] dropping all tables for a clean test run.")
-        drop_all_tables(conn)
-    except Exception as e:
-        print(f"[main] error dropping tables: {e}")
-        conn.close()
-        return
-
-    # prompt user for main table name
-    default_table = os.path.splitext(os.path.basename(filename))[0]
-    user_table = input(f"enter main table name (default: {default_table}): ").strip()
-    main_table_name = user_table if user_table else default_table
-
-    # inject dataset
-    print(f"[main] injecting dataset from {filename} into database {db_path}.")
-    success = inject_dataset(conn, filename, main_table_name)
-    print(f"[main] data injection successful?: {success}")
-
-    if not success:
-        conn.close()
-        return
-
-    # menu options for CLI interaction
-    while success:
-        print("\nchoose an option:")
-        print("1. list all tables")
-        print("2. query a specific table")
-        print("3. query main table")
-        print("4. database summary")
-        print("5. quit")
-        choice = input("[main] enter option number: ").strip()
-        if choice == "1":
-            cursor = conn.cursor()
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
-            tables = [row[0] for row in cursor.fetchall()]
-            print("[main] tables in database:", tables)
-        elif choice == "2":
-            table_name = input("[main] enter table name: ").strip()
-            cursor = conn.cursor()
-            try:
-                cursor.execute(f'SELECT * FROM "{table_name}"')
-                rows = cursor.fetchall()
-                if not rows:
-                    print(f"[main] no data found in table '{table_name}'.")
-                else:
-                    print(f"[main] rows in '{table_name}':")
-                    for row in rows:
-                        print(row)
-            except Exception as e:
-                print(f"[main] error querying table '{table_name}': {e}")
-        elif choice == "3":
-            cursor = conn.cursor()
-            print(f"[main] listing all rows in main table '{main_table_name}'.")
-            try:
-                cursor.execute(f'SELECT * FROM "{main_table_name}";')
-                rows = cursor.fetchall()
-                if not rows:
-                    print(f"[main] no data found in {main_table_name} table.")
-                else:
-                    print(f"[main] rows in {main_table_name} table:")
-                    for row in rows:
-                        print(row)
-            except Exception as e:
-                print(f"[main] error querying {main_table_name} table: {e}")
-        elif choice == "4":
-            print("[main] printing database summary.")
-            print_database_summary(conn, main_table_name)
-        elif choice == "5":
-            print("[main] exiting.")
-            break
-        else:
-            print("[main] invalid option. Try again.")
-
-    conn.close()
-
-if __name__ == "__main__":
-    main()
